Tugas Pertemuan 6
Index:
- Menghandle jika data kosong

Store:
-Menghandle jika salah satu data tidak dikirim ketika menambahkan data

Update:
- Menghandle jika resource yang diminta tidak ada
- Menghandle jika salah satu data tidak dikirim ketika meng-update data

Delete:
- Menghandle jika resource yang diminta tidak ada

Add authentication (sanctum)
- Protect routes
- Login Authentication with token

+++ laravel-student/app/Http/Controllers/StudentController.php
@@ -4,20 +4,52 @@ namespace App\Http\Controllers;
 
 use Illuminate\Http\Request;
 use App\Models\Student;
+use App\Http\Requests\students\StoreRequest;
+use App\Http\Requests\students\UpdateRequest;
 
 class StudentController extends Controller
 {
     public function index(Student $student)
     {
-        return response()->json([
-            'success' => true,
-            'message' => 'Daftar data mahasiswa',
-            'data' => $student->all()
-        ], 200);
+        if ($student->all()) {
+            return response()->json([
+                'success' => true,
+                'message' => 'Daftar data mahasiswa',
+                'data' => $student->all()
+            ], 200);
+        } else {
+            return response()->json([
+                'success' => false,
+                'message' => 'Daftar data mahasiswa',
+                'data' => []
+            ], 404);
+        }
     }
 
-    public function store(Student $student)
+    public function show(Student $student, $id)
     {
+
+        $student = $student->find($id);
+
+        if ($student) {
+            return response()->json([
+                'success' => true,
+                'message' => 'Detail data mahasiswa',
+                'data' => $student
+            ], 200);
+        } else {
+            return response()->json([
+                'success' => false,
+                'message' => 'Data mahasiswa tidak ditemukan',
+                'data' => ''
+            ], 404);
+        }
+    }
+
+    public function store(Student $student, StoreRequest $storeStudentRequest,)
+    {
+        $storeStudentRequest->validated();
+
         $student::create([
             'nama' => request()->nama,
             'nim' => request()->nim,
@@ -32,33 +64,52 @@ class StudentController extends Controller
         ], 200);
     }
 
-    public function update($id, Request $request, Student $student)
+    public function update(Student $student, UpdateRequest $updateRequest, $id)
     {
 
-        $student = $student->find($id);
-        $student->update([
-            'nama' => request()->nama,
-            'nim' => request()->nim,
-            'jurusan' => request()->jurusan,
-            'email' => request()->email
-        ]);
+        $updateRequest->validated();
 
-        return response()->json([
-            'success' => true,
-            'message' => 'Data mahasiswa berhasil diupdate',
-            'data' => $student
-        ], 200);
+        $student = $student->find($id);
+
+        if ($student) {
+            $student->update([
+                'nama' => request()->nama,
+                'nim' => request()->nim,
+                'jurusan' => request()->jurusan,
+                'email' => request()->email
+            ]);
+
+            return response()->json([
+                'success' => true,
+                'message' => 'Data mahasiswa berhasil diupdate',
+                'data' => $student
+            ], 200);
+        } else {
+            return response()->json([
+                'success' => false,
+                'message' => 'Data mahasiswa tidak ditemukan',
+                'data' => ''
+            ], 404);
+        }
     }
 
     public function destroy($id, Student $student)
     {
-        $student = $student->find($id);
-        $student->delete();
+        if ($student) {
+            $student = $student->find($id);
+            $student->delete();
 
-        return response()->json([
-            'success' => true,
-            'message' => 'Data mahasiswa berhasil dihapus',
-            'data' => $student
-        ], 200);
+            return response()->json([
+                'success' => true,
+                'message' => 'Data mahasiswa berhasil dihapus',
+                'data' => $student
+            ], 200);
+        } else {
+            return response()->json([
+                'success' => false,
+                'message' => 'Data mahasiswa tidak ditemukan',
+                'data' => ''
+            ], 404);
+        }
     }
 }
